[Infra/Feat] Grafana, Loki, Promtail 로그 수집 및 시각화 구성 추가 S12P31D103-133

See merge request s12-final/S12P31D103!132

+++ Android/data/src/main/java/com/hogumiwarts/data/entity/remote/Response/light/GetLightStatusResponse.kt
@@ -11,6 +11,6 @@ data class GetLightStatusResponse(
     val activated: Boolean,
     val brightness: Int,
     val lightTemperature: Int,
-    val hue: Int,
+    val hue: Float,
     val saturation: Float,
 )
